FlightRisk/requirements.txt
requests
python-dotenv
pandas
openpyxl
google-cloud-bigquery
google-cloud-bigquery-storage
pandas-gbq>=0.26.1
pyarrow
db-dtypes
